+++ examples/snl/bsp.c
@@ -11,16 +11,20 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+// BSP for Sense and Locate board
+
 #include "bsp.h"
 #include <string.h>
 #include "myriota_hardware_api.h"
 #include "myriota_user_api.h"
 
-#define LED_PIN PIN_GPIO3
+#define GREEN_LED_GPIO PIN_GPIO8
+#define RED_LED_GPIO PIN_GPIO7
+
 #define GNSS_EN_PIN PIN_GPIO4
 #define ANT_SEL_PIN PIN_GPIO6
-#define BATTERY_MEASUREMENT_EN_PIN PIN_GPIO2
 #define BATTERY_MEASUREMENT_ADC_PIN PIN_ADC1
+#define BATTERY_MEASUREMENT_EN_PIN PIN_GPIO2
 #define DEBUG_INTERFACE UART_0
 #define DEBUG_BAUDRATE 115200
 #define MODULE_BAND_PIN PIN_BAND
@@ -42,8 +46,8 @@ __attribute__((weak)) char *BoardEnvGet() {
 
 int BoardInit(void) {
   BoardLEDDeinit();
-  // Avoid current leak from the battery measurement switch if floating
-  GPIOSetModeInput(BATTERY_MEASUREMENT_EN_PIN, GPIO_PULL_DOWN);
+  // Avoid current leak from the battery measurement switch when floating
+  GPIOSetModeInput(PIN_GPIO2, GPIO_PULL_DOWN);
   return 0;
 }
 
@@ -52,9 +56,8 @@ __attribute__((weak)) int BoardStart(void) {
   Delay(200);
   LedTurnOff();
   GPIOSetModeInput(MODULE_BAND_PIN, GPIO_NO_PULL);
-  printf("Myriota development board %s variant %s %s\n",
-         GPIOGet(MODULE_BAND_PIN) == GPIO_HIGH ? "VHF" : "UHF", ModuleIDGet(),
-         RegistrationCodeGet());
+  printf("Myriota sense and locate board %s variant %s\n",
+         GPIOGet(MODULE_BAND_PIN) == GPIO_HIGH ? "VHF" : "UHF", ModuleIDGet());
   char *EnvStr = BoardEnvGet();
   if (EnvStr && strlen(EnvStr)) {
     printf("Using env %s\n", EnvStr);
@@ -62,11 +65,7 @@ __attribute__((weak)) int BoardStart(void) {
   return 0;
 }
 
-// If battery measurement is not supported, both voltage and the return
-// value should be 0.
 int BoardBatteryVoltGet(uint32_t *mv) {
-// Compensate for the voltage drop across the diode Q202
-#define VEXT_COMPENSATE 15    // [mV]
 #define SWITCH_SETTLE_TIME 1  // [ms]
 #define AVERAGE_COUNT 3
 
@@ -89,24 +88,34 @@ int BoardBatteryVoltGet(uint32_t *mv) {
   }
   GPIOSetModeInput(ControlPin, GPIO_PULL_DOWN);
   *mv = volt / AVERAGE_COUNT;
-  *mv *= 2;  // Divider on the development board
-  *mv += VEXT_COMPENSATE;
+  *mv *= 2;  // Divider on board
   return 0;
 }
 
 void BoardLEDInit(void) {
-  GPIOSetModeOutput(LED_PIN);
-  GPIOSetLow(LED_PIN);
+  GPIOSetModeOutput(GREEN_LED_GPIO);
+  GPIOSetLow(GREEN_LED_GPIO);
+  GPIOSetModeOutput(RED_LED_GPIO);
+  GPIOSetLow(RED_LED_GPIO);
 }
 
-void BoardLEDDeinit(void) { GPIOSetModeInput(LED_PIN, GPIO_PULL_DOWN); }
+void BoardLEDDeinit(void) {
+  GPIOSetModeInput(GREEN_LED_GPIO, GPIO_PULL_DOWN);
+  GPIOSetModeInput(RED_LED_GPIO, GPIO_PULL_DOWN);
+}
 
-void BoardLEDTurnOn(void) { GPIOSetHigh(LED_PIN); }
+void BoardLEDTurnOn(void) {
+  GPIOSetHigh(GREEN_LED_GPIO);
+  GPIOSetHigh(RED_LED_GPIO);
+}
 
-void BoardLEDTurnOff(void) { GPIOSetLow(LED_PIN); }
+void BoardLEDTurnOff(void) {
+  GPIOSetLow(GREEN_LED_GPIO);
+  GPIOSetLow(RED_LED_GPIO);
+}
 
 void BoardLEDToggle(void) {
-  if (GPIO_LOW == GPIOGet(LED_PIN))
+  if (GPIO_LOW == GPIOGet(GREEN_LED_GPIO))
     BoardLEDTurnOn();
   else
     BoardLEDTurnOff();
@@ -124,9 +133,9 @@ void BoardGNSSPowerDisable(void) {
 
 bool BoardGNSSPowerIsEnabled(void) { return GPIOGet(GNSS_EN_PIN); }
 
-__attribute__((weak)) int BoardAntennaSelect(RadioMode Mode, RadioBand Band) {
-  bool OnBoardAntenna = true;
-  if (Band == RADIO_BAND_VHF) OnBoardAntenna = false;
+int BoardAntennaSelect(RadioMode Mode, RadioBand Band) {
+  bool IsUHFAntenna = true;
+  if (Band == RADIO_BAND_VHF) IsUHFAntenna = false;
 
   GPIOSetModeOutput(ANT_SEL_PIN);
 
@@ -136,7 +145,7 @@ __attribute__((weak)) int BoardAntennaSelect(RadioMode Mode, RadioBand Band) {
     return 0;
   }
 
-  if (OnBoardAntenna) {
+  if (IsUHFAntenna) {
     GPIOSetHigh(ANT_SEL_PIN);
   } else {
     GPIOSetLow(ANT_SEL_PIN);
@@ -165,3 +174,4 @@ int BoardDebugRead(uint8_t *Rx, size_t Length) {
 __attribute__((weak)) void BoardSleepEnter(void) {}
 
 __attribute__((weak)) void BoardSleepExit(void) {}
+